requirements.txt
pandas==2.1.4
netmiko==4.2.0
openpyxl==3.1.2
paramiko==3.4.0
cryptography==41.0.7
msoffcrypto-tool==5.0.0
xlrd==2.0.1
colorama==0.4.6
